Refactor out (part of) errors and util to edb.server.ext
