+++ edb/server/ext/errors.py
@@ -1,0 +1,42 @@
+#
+# This source file is part of the EdgeDB open source project.
+#
+# Copyright 2023-present MagicStack Inc. and the EdgeDB authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+class ExtError(Exception):
+    """Base class for all exceptions raised by extension code."""
+
+    pass
+
+
+class MissingConfiguration(ExtError):
+    """Required configuration is missing."""
+
+    def __init__(self, key: str, description: str):
+        self.key = key
+        self.description = description
+
+    def __repr__(self):
+        return (
+            f"{self.__class__.__name__}("
+            f"key={self.key!r} "
+            f"description={self.description!r}"
+            ")"
+        )
+
+    def __str__(self) -> str:
+        return f"{self.description}: {self.key}"
