+++ edb/server/ext/util.py
@@ -20,6 +20,7 @@
 from typing import TypeVar, Type, Mapping, overload, Any
 
 from edb.server.config.ops import SettingValue
+
 from . import errors
 
 T = TypeVar("T")
