+++ edb/server/protocol/auth_ext/errors.py
@@ -16,8 +16,10 @@
 # limitations under the License.
 #
 
+from edb.server.ext.errors import ExtError
 
-class AuthExtError(Exception):
+
+class AuthExtError(ExtError):
     """Base class for all exceptions raised by the auth extension."""
 
     pass
@@ -38,25 +40,6 @@ class NotFound(AuthExtError):
 
     def __str__(self) -> str:
         return self.description
-
-
-class MissingConfiguration(AuthExtError):
-    """Required configuration is missing."""
-
-    def __init__(self, key: str, description: str):
-        self.key = key
-        self.description = description
-
-    def __repr__(self):
-        return (
-            f"{self.__class__.__name__}("
-            f"key={self.key!r} "
-            f"description={self.description!r}"
-            ")"
-        )
-
-    def __str__(self) -> str:
-        return f"{self.description}: {self.key}"
 
 
 class InvalidData(AuthExtError):
