+++ edb/server/protocol/auth_ext/oauth.py
@@ -20,9 +20,10 @@
 import json
 
 from typing import Any
+from edb.server.ext import util
 from edb.server.protocol import execute
 
-from . import errors, util, data, base, http_client
+from . import errors, data, base, http_client
 
 
 class Client:
